# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Transfert Bancaire",
        "description": "API pour la gestion des comptes bancaires",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "Serveur de développement"
        },
        {
            "url": "https://api-transfert.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/v1/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister tous les comptes avec filtres optionnels",
                "description": "Récupère une liste paginée des comptes bancaires avec filtrage optionnel par numéro de compte, nom d'utilisateur, type et statut",
                "operationId": "c09c55b936e3520140768760edfa911e",
                "parameters": [
                    {
                        "name": "numero_compte",
                        "in": "query",
                        "description": "Filtrer par numéro de compte (correspondance partielle)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nom_user",
                        "in": "query",
                        "description": "Filtrer par nom d'utilisateur (correspondance partielle)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de compte",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "epargne",
                                "cheque"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut du compte",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "bloque"
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "dateCreation",
                            "enum": [
                                "dateCreation",
                                "numero_compte",
                                "type_compte",
                                "status_compte"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comptes récupérés avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Compte"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte",
                "description": "Créer un nouveau compte bancaire. Si user_id est fourni, utilise le client existant. Sinon, crée un nouveau client avec les informations fournies.",
                "operationId": "aa399240cc7c1f272aa92fe6d6b4e162",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero_compte",
                                    "type_compte",
                                    "telephone"
                                ],
                                "properties": {
                                    "numero_compte": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "type_compte": {
                                        "type": "string",
                                        "enum": [
                                            "epargne",
                                            "cheque"
                                        ],
                                        "example": "epargne"
                                    },
                                    "status_compte": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "bloque"
                                        ],
                                        "example": "active"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "+221771234567"
                                    },
                                    "user_id": {
                                        "description": "ID utilisateur existant (optionnel)",
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "client_name": {
                                        "description": "Nom du client (requis si pas de user_id)",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "client_email": {
                                        "description": "Email du client (requis si pas de user_id)",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "client_password": {
                                        "description": "Mot de passe du client (requis si pas de user_id)",
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte créé avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Compte": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "numero_compte": {
                        "type": "string",
                        "example": "123456789"
                    },
                    "type_compte": {
                        "type": "string",
                        "enum": [
                            "epargne",
                            "cheque"
                        ],
                        "example": "epargne"
                    },
                    "status_compte": {
                        "type": "string",
                        "enum": [
                            "active",
                            "bloque"
                        ],
                        "example": "active"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+221771234567"
                    },
                    "client": {
                        "$ref": "#/components/schemas/Client"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Client": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "client",
                            "admin"
                        ],
                        "example": "client"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
    "security": [
        {
            "passport": []
        }
    ]
}